kteam-mbox: do not use a lock file
The lock files are kind of confusing because they stick around after
we're done. We don't really need them so use a try/finally pattern. The
only reason I used a lock file to begin with was for the guarnteed close
side effect which can be more cleanly achieved with try/finally.

Signed-off-by: Cory Todd <cory.todd@canonical.com>