patch_set: add classifier back to PatchSet
There are two ways to classify an email. Locally, meaning without
context of the thread, and in the context of the thread. We should never
leave a message uncategorized, that would be surprising. We will now
return messages that are at a minimum, locally classified. PatchSet can
now classify within the context of a thread.

Signed-off-by: Cory Todd <cory.todd@canonical.com>