snap: promote to stable
Cleanup the messy pyproject and remove the ugly setup.py. The former
now holds all metadata and the requirements are all contained within
the requirements.txt file. This makes it easy to install the packages
for local dev without having to maintain a file list in three places.

Remove the tools/version.py script that doesn't seem to work. Instead, we will set the version in ml_check.init.py. This is more
of a standard approach anyways.

Signed-off-by: Cory Todd <cory.todd@canonical.com>